Refactor Farming Hud
